Zomato/architecture/phase_5_response_delivery/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
groq>=0.9.0
